+++ app/jobs/provider/pc_create_or_update_job.rb
@@ -4,6 +4,6 @@ class Provider::PcCreateOrUpdateJob < ApplicationJob
   queue_as :pc_subscriber
 
   def perform(provider, is_active, modified_at)
-    Provider::PcCreateOrUpdate.new(provider, is_active, modified_at).call
+    Provider::PcCreateOrUpdate.call(provider, is_active, modified_at)
   end
 end