+++ spec/lib/import/resources_spec.rb
@@ -59,7 +59,7 @@ describe Import::Resources do
   def expect_responses(test_url, services_response = nil)
     unless services_response.nil?
       allow_any_instance_of(Faraday::Connection).to receive(:get)
-        .with("#{test_url}/service/adminPage/all?quantity=10000&from=0")
+        .with("#{test_url}/public/resource/adminPage/all?quantity=10000&from=0")
         .and_return(services_response)
     end
   end